pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurosync-pro"
version = "0.1.0"
description = "NeuroSync Pro — EEG, audio, MVP (scaffold)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "bleak>=0.21.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
gui = [
    "PySide6>=6.5.0",
]

[project.scripts]
neurosync-pro = "neurosync_pro.cli:main"

[project.urls]
Homepage = "https://github.com/useitone/LLM_CHAT"
Repository = "https://github.com/useitone/LLM_CHAT"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
